save point

+++ db/migrate/20190205184352_add_company_to_user.rb
@@ -1,5 +1,5 @@
 class AddCompanyToUser < ActiveRecord::Migration[5.2]
   def change
-    add_column :users, :company_id, :bigint
+    add_column :users, :company_id, :integer
   end
 end
